requirements.txt
antlr4-python3-runtime
antlr4-tools
black
cfpq-data
networkx
numpy
pre-commit
prettytable
pydot
pyformlang
pytest
scipy
